% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_effect_qc_setup.R
\name{batch_effect_qc_setup}
\alias{batch_effect_qc_setup}
\title{set up the QC notebook with the data sets which are used throughout}
\usage{
batch_effect_qc_setup(
  dds,
  percent_samples_for_expr_fltr = 0.015,
  cpm_count_thres = 3
)
}
\arguments{
\item{dds}{a deseq data set object}

\item{percent_samples_for_expr_fltr}{this is used to set a minimum number of
samples which have more than a certain number of cpm count. ie
num_samples = floor(ncol(dds)*percent_samples_for_expr_fltr)}

\item{cpm_count_thres}{the number of cpm required for a given gene to pass
in a given sample. ie expr_fltr =
rowSums(cpm(counts(dds)) > cpm_count_thres ) >= num_samples}
}
\value{
a list with the various vst data and dds data objects. This is
very large -- need a lot of RAM.
}
\description{
set up the QC notebook with the data sets which are used throughout
}
