% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/before_after_pca_and_plate_regression_plots.R
\name{before_after_pca_and_plate_regression_plots}
\alias{before_after_pca_and_plate_regression_plots}
\title{Before and After PCA and Plate Regression Plots}
\usage{
before_after_pca_and_plate_regression_plots(
  counts,
  pca_basis_data,
  x_pc,
  y_pc,
  meta,
  plate_colors,
  plot_title,
  pc_levels = NA
)
}
\arguments{
\item{counts}{log normalized "counts"}

\item{pca_basis_data}{data from which to create the PCA basis vectors}

\item{x_pc}{which PC to plot on the x axis}

\item{y_pc}{which PC to plot on the Y axis}

\item{meta}{sample metadata. Rows should be in same order as counts}

\item{plate_colors}{vector of colors for each batch}

\item{plot_title}{title of plot}

\item{pc_levels}{a vector of PC levels, eg 1,3,6,2,4,5 which will control
the order which the PCs are plotted}
}
\value{
A list that includes plots and data
}
\description{
Show the effect of regressing out plate effects on the PCs
}
