% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_txi_objects.R
\name{combine_txi_objects}
\alias{combine_txi_objects}
\title{combine two txi objects}
\usage{
combine_txi_objects(txi1, txi2)
}
\arguments{
\item{txi1}{one of the two txi objects to combine}

\item{txi2}{the other txi object}
}
\value{
the result of combining \code{txi1} and \code{txi2}
}
\description{
combine two txi objects
}
\note{
the column names of \code{txi1} and \code{txi2} are required to be unique
}
