% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction_change_in_percent_intergenic.R
\name{direction_change_in_percent_intergenic}
\alias{direction_change_in_percent_intergenic}
\title{make a plot showing the direction of percent intergenic on PC1/2 axis}
\usage{
direction_change_in_percent_intergenic(
  pca_df_list,
  plate_colors,
  slope_dir = 1
)
}
\arguments{
\item{pca_df_list}{list of PCA dfs}

\item{plate_colors}{a vector corresponding to plate colors. should be equal
in length to the number of plates}

\item{slope_dir}{either 1 or -1, to flip slope depending on how the PCs are
calculated}
}
\value{
A plot showing the line which results from conducting a linear
module with formula percent_intergenic ~ PC1 + PC2
}
\description{
make a plot showing the direction of percent intergenic on PC1/2 axis
}
