% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dds_restimate_sizeFactors.R
\name{filter_dds_restimate_sizeFactors}
\alias{filter_dds_restimate_sizeFactors}
\title{Apply boolean vector filtering on the columns and/or rows and return
filtered DDS object with size factors}
\usage{
filter_dds_restimate_sizeFactors(dds, gene_fltr = NULL, qc_sample_fltr = NULL)
}
\arguments{
\item{dds}{a DeseqDataSet object}

\item{gene_fltr}{a boolean vector of length nrow(dds) where TRUE are the
indicies to keep. Optional}

\item{qc_sample_fltr}{a boolean vector of length ncol(dds) where TRUE are
the indicies to keep. Optional}
}
\value{
a filtered DESeqDataSet. Size factors are calculated.
}
\description{
Apply boolean vector filtering on the columns and/or rows and return
filtered DDS object with size factors
}
