% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_all_by_all_compiled_results.R
\name{parse_all_by_all_compiled_results}
\alias{parse_all_by_all_compiled_results}
\title{summarize an all_by_all_compiled_metrics dataframe for upload to the
database}
\usage{
parse_all_by_all_compiled_results(all_by_all_df)
}
\arguments{
\item{all_by_all_df}{a wgs all by all compiled metrics dataframe with the
columns 'total_variants', 'library_id', 'rna_subject', 'dna_subject',
'total_variants', 'matching_variants', 'match_ratio'}
}
\value{
a dataframe suitable for upload to the full_wgs_compare table
}
\description{
summarize an all_by_all_compiled_metrics dataframe for upload to the
database
}
