% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pca_sex_qc}
\alias{pca_sex_qc}
\title{PCA Sex expression plots}
\format{
An object of class \code{list} of length 4.
}
\source{
vignettes/sex-label-check
}
\usage{
pca_sex_qc
}
\description{
This is an example of the two PCA plots, and the accompanying tables,
which are used to do the PCA sex labelling QC check
}
\keyword{datasets}
