% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_counts_onto_orig_pcs.R
\name{project_counts_onto_orig_pcs}
\alias{project_counts_onto_orig_pcs}
\title{Project counts, presumably with some effect removed, onto different
PCs}
\usage{
project_counts_onto_orig_pcs(dds, effect_removed_counts)
}
\arguments{
\item{dds}{a DESeqDataSet object}

\item{effect_removed_counts}{A "count" matrix where some effect
(coefficient) has been removed}
}
\value{
the reult of projecting the effect removed counts onto the old PCs.
This should show reduced variance/tighter expected grouping
}
\description{
Project counts, presumably with some effect removed, onto different
PCs
}
