% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_parameter_effects.R
\name{remove_parameter_effects}
\alias{remove_parameter_effects}
\title{remove coefficient effects from counts}
\usage{
remove_parameter_effects(deseq_object, col_indicies)
}
\arguments{
\item{deseq_object}{a deseq data object REQUIRED: the object must have
been created with a model.matrix rather than a formula for the
design argument}

\item{col_indicies}{a numeric vector corresponding to the column
indicies of the batch parameters you'd like to remove}
}
\value{
a log2 scale gene by samples matrix with desired effects removed
}
\description{
subtract effect from norm counts of a single factor from
coef x design. coef is in normalized log space. dds must have been
created with model.matrix
}
\note{
works for both formula and model.matrix designs in the dds object
}
