% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_mislabel_setup.R
\name{sex_mislabel_setup}
\alias{sex_mislabel_setup}
\title{set up the QC notebook with the data sets which are used throughout}
\usage{
sex_mislabel_setup(
  dds_rds_path,
  percent_samples_for_expr_fltr = 0.015,
  cpm_count_thres = 3
)
}
\arguments{
\item{dds_rds_path}{path to a dds object, either gene or transcript counts}

\item{percent_samples_for_expr_fltr}{this is used to set a minimum number of
samples which have more than a certain number of cpm count. ie
num_samples = floor(ncol(dds)*percent_samples_for_expr_fltr)}

\item{cpm_count_thres}{the number of cpm required for a given gene to pass
in a given sample. ie expr_fltr =
rowSums(cpm(counts(dds)) > cpm_count_thres ) >= num_samples}
}
\value{
a list which contains the raw, expression filtered and expression
filtered and passing sets
}
\description{
set up the QC notebook with the data sets which are used throughout
}
